credentials: inline private methods (#97)
* credentials: inline __validate_tokens

This is a single-use method, and inlining
it ensures NoTokenException cannot happen,
so this saves almost 20 lines of code while
preserving the existing functionality.

Eliminating NoTokenException also removes
the need to test this error scenario.

* credentials: simplify __ensure_tokens

Initialize the expire times to the call
time of the constructor instead of None.
This allows the first if-statement in
__ensure_tokens to be removed since
the following time comparisons will
call __exchange_credentials.

* credentials: inline single-use methods

Set the data dynamically in __ensure_tokens,
and make a single call to __token_exchange
with that data. This saves another ~10
lines of code and reduces the surface
area of the Credentials object.