query: check return value of get
The request might not succeed, in
which case the json() call will
give a mysterious:

json.decoder.JSONDecodeError: Expecting value: line 1 column 1 (char 0)

Do a .raise_for_status() before
calling json() on the response.
This also throws an exception
in the face of the user,
but it's easier to understand
that something went wrong
with the HTTP request.