Use Ruff instead of black
This replaces black with Ruff.
The coding style of Ruff is based
on black so there are no significant
differences as evidenced by the size
of the diff of this PR. The main
difference is that Ruff is significantly
faster.

Ruff also provides the functionality
of isort and Flake8 with various plugins,
but this PR just replaces the formatting
parts done by black, any further linting
can be done in a future PR.