query_feature: retry on HTTP errors (#183)
* query: add logger to FeatureQuery

Make a log object available so diagnostic information
can be emitted.

* query_feature: retry on HTTP errors

Implement the retry-logic from
download_feature in query_feature
so CDSETool retries a few times
when there is a HTTP error, and
does not crash and give a backtrace
back to the user.

* query_feature: retry on non-200 response

Since there is now a retry-loop
in query_feature, use that to retry
when there is a non-200 response
from the server. This avoids giving
a backtrace to the end-user for errors
that are more or less expected.