query: permit skipping term validation
Add an option to permit skipping term
validation that is disabled by default.
Not validating search terms reduces
the number of calls to remote servers,
which is useful for scripted environments
where the search terms are known to
be valid.

Without search term validation, invalid input
will cause errors when contacting the remote
servers.