query: retry even more (#207)
* query: perform error check early

Get the value for the key from the description
and check that it is not None before validating
the value. This avoids handling None inside
the validation code.

* query: make type more specific

The object type is more specific than
Any, so use that.

* query: add retry-loop to _get_describe_doc

Add a loop to retry on failure. Old logic
is still present so this makes no functional
difference.

* query: sleep+retry on server failure

Add the sleep-statement from the download
code to the _get_describe_doc function
as well.

* credentials: increase retries

The Copernicus server fail intermittently,
so try more times to get the response.

* query: permit skipping term validation

Add an option to permit skipping term
validation that is disabled by default.
Not validating search terms reduces
the number of calls to remote servers,
which is useful for scripted environments
where the search terms are known to
be valid.

Without search term validation, invalid input
will cause errors when contacting the remote
servers.