pyproject: add dependency-groups.dev for Ruff (#237)
The Ruff action looks for Ruff version
in dependency-gruops.dev, so make
such a group for Ruff.

This prevents pylint upgrades from
failing because the most recent
Ruff changed something in the formatting.

Keep the other development dependencies
in their previous section since
pip has not yet implemented the support
for dependency-groups.