Delay importing large packages
The import time for cdsetool.download
is 626723 us, and geopandas is
responsible for 622825 us of that.

Push imports of various packages
that have reasonably long load
times into the function where
they are used.

This improves CLI performance
for us, and also benefits any
downsteam users that build
their own CLI around CDSETool.